change REM and REMU back to 64 bit because test-shift failed when 32bi

+++ src/sst/elements/vanadis/decoder/vriscv64decoder.h
@@ -735,7 +735,7 @@ protected:
                         output->verbose(
                             CALL_INFO, 16, 0, "-----> REM %" PRIu16 " <- %" PRIu16 " %% %" PRIu16 "\n", rd, rs1, rs2);
                         bundle->addInstruction(
-                            new VanadisModuloInstruction<int32_t>(
+                            new VanadisModuloInstruction<int64_t>(
                                 ins_address, hw_thr, options, rd, rs1, rs2));
                         decode_fault = false;
                     } break;
@@ -758,7 +758,7 @@ protected:
                         output->verbose(
                             CALL_INFO, 16, 0, "-----> REMU %" PRIu16 " <- %" PRIu16 " %% %" PRIu16 "\n", rd, rs1, rs2);
                         bundle->addInstruction(
-                            new VanadisModuloInstruction<uint32_t>(
+                            new VanadisModuloInstruction<uint64_t>(
                                 ins_address, hw_thr, options, rd, rs1, rs2));
                         decode_fault = false;
                     } break;
